requirements.txt
# LLM Framework
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0

# Vector Database
chromadb>=0.4.22

# Embeddings
sentence-transformers>=2.2.2

# LLM (for Ollama)
ollama>=0.1.6

# Web UI
streamlit>=1.29.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0

# HTTP requests (for price comparison APIs)
requests>=2.31.0
httpx>=0.25.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0